gemini_tts_tool/core/client.py
"""Gemini client management for TTS operations.

Note: This code was generated with assistance from AI coding tools
and has been reviewed and tested by a human.
"""

import os

from google import genai


class GeminiClientError(Exception):
    """Base exception for Gemini client errors."""

    pass


class AuthenticationError(GeminiClientError):
    """Raised when authentication fails."""

    pass


def create_client(
    api_key: str | None = None,
    use_vertex: bool = False,
    project: str | None = None,
    location: str | None = None,
) -> genai.Client:
    """Create and configure a Gemini client for TTS operations.

    Supports both Gemini Developer API and Vertex AI authentication.

    Args:
        api_key: API key for Gemini Developer API (optional, reads from env)
        use_vertex: Whether to use Vertex AI instead of Developer API
        project: Google Cloud project ID (required for Vertex AI)
        location: Google Cloud location (required for Vertex AI)

    Returns:
        Configured Gemini client

    Raises:
        AuthenticationError: If authentication configuration is invalid
    """
    # Vertex AI configuration
    # Check if Vertex AI is enabled via environment or parameter
    use_vertex_env = os.getenv("GOOGLE_GENAI_USE_VERTEXAI", "").lower() in ("true", "1", "yes")
    if use_vertex or use_vertex_env:
        # Read from parameters or environment
        if not project:
            project = os.getenv("GOOGLE_CLOUD_PROJECT")
        if not location:
            location = os.getenv("GOOGLE_CLOUD_LOCATION")

        if not project:
            raise AuthenticationError(
                "GOOGLE_CLOUD_PROJECT environment variable or --project is required for Vertex AI"
            )
        if not location:
            raise AuthenticationError(
                "GOOGLE_CLOUD_LOCATION environment variable or --location is required for Vertex AI"
            )
        try:
            return genai.Client(vertexai=True, project=project, location=location)
        except Exception as e:
            raise AuthenticationError(f"Failed to create Vertex AI client: {e}") from e

    # Gemini Developer API configuration
    # Priority: 1. Explicit api_key param, 2. GOOGLE_API_KEY, 3. GEMINI_API_KEY
    if not api_key:
        api_key = os.getenv("GOOGLE_API_KEY") or os.getenv("GEMINI_API_KEY")

    if not api_key:
        raise AuthenticationError(
            "API key is required. Set GEMINI_API_KEY or GOOGLE_API_KEY environment variable, "
            "or use --api-key option. "
            "Get your API key from https://aistudio.google.com/app/apikey"
        )

    try:
        return genai.Client(api_key=api_key)
    except Exception as e:
        raise AuthenticationError(f"Failed to create Gemini client: {e}") from e


def validate_client(client: genai.Client) -> None:
    """Validate that the client is properly configured.

    Args:
        client: Gemini client to validate

    Raises:
        GeminiClientError: If client validation fails
    """
    if not isinstance(client, genai.Client):
        raise GeminiClientError(f"Invalid client type: {type(client)}")
